+++ src/common.c
@@ -5,9 +5,9 @@
 #include <time.h>
 #include <string.h>
 
-void report_time(int dt,int nt,int base)
+void report_time(double dt,int nt,double base)
 {
-  printf("%f %d\n",((float) dt)/CLOCKS_PER_SEC/nt*1000,dt/base);
+  printf("%f %f\n",((double) dt)/CLOCKS_PER_SEC/nt*1000,dt/base);
 }
 
 void check_ciphertext(byte *out,byte *outex,int nbyte)
@@ -19,17 +19,17 @@ void check_ciphertext(byte *out,byte *outex,int nbyte)
   }
 }
 
-int runalgo(void (*algo)(byte *,byte *,byte *),byte *in,byte *out,byte *key,byte *outex,int nbyte,int nt,int base)
+double runalgo(void (*algo)(byte *,byte *,byte *),byte *in,byte *out,byte *key,byte *outex,int nbyte,int nt,double base)
 {
   int i;
   clock_t start,end;
-  int dt;
+  double dt;
   start=clock();
   for(i=0;i<nt;i++)
     algo(in,out,key);
   end=clock();
-  dt=(int) (end-start);
-  if (base==0) base=dt;
+  dt=(double) (end-start);
+  if (base==0.0f) base=dt;
   report_time(dt,nt,base);
   check_ciphertext(out,outex,nbyte);
   return dt;
